Revert "Merge pull request #1 from CHOMPStation2/master"
This reverts commit b85a3b312d45ce4a323475f702261c783391ac28, reversing
changes made to a13aa8f3679ae94815291e7570926e28135e8dbe.

+++ code/game/objects/items/devices/megaphone.dm
@@ -10,52 +10,53 @@
 	var/list/insultmsg = list("FUCK EVERYONE!", "I'M A TERRORIST!", "ALL SECURITY TO SHOOT ME ON SIGHT!", "I HAVE A BOMB!", "CAPTAIN IS A COMDOM!", "GLORY TO ALMACH!")
 
 /obj/item/device/megaphone/proc/can_broadcast(var/mob/living/user)
-	if(user.client)
+	if (user.client)
 		if(user.client.prefs.muted & MUTE_IC)
 			to_chat(user, "<span class='warning'>You cannot speak in IC (muted).</span>")
-			return FALSE
+			return 0
 	if(!(ishuman(user) || user.isSynthetic()))
 		to_chat(user, "<span class='warning'>You don't know how to use this!</span>")
-		return FALSE
+		return 0
 	if(user.silent)
-		return FALSE
-	if(spamcheck > world.time)
-		to_chat(user, "<span class='warning'>[src] needs to recharge!</span>")
-		return FALSE
-	if(loc != user)
-		return FALSE
-	if(user.stat != CONSCIOUS)
-		return FALSE
-	return TRUE
+		return 0
+	if(spamcheck)
+		to_chat(user, "<span class='warning'>\The [src] needs to recharge!</span>")
+		return 0
+	return 1
 
 /obj/item/device/megaphone/proc/do_broadcast(var/mob/living/user, var/message)
-	if(emagged)
-		if(insults)
-			user.audible_message("<B>[user.GetVoice()]</B>[user.GetAltName()] broadcasts, <FONT size=3>\"[pick(insultmsg)]\"</FONT>")
-			insults--
+	if ((src.loc == user && usr.stat == 0))
+		if(emagged)
+			if(insults)
+				user.audible_message("<B>[user]</B> broadcasts, <FONT size=3>\"[pick(insultmsg)]\"</FONT>")
+				insults--
+			else
+				to_chat(user, "<span class='warning'>*BZZZZzzzzzt*</span>")
 		else
-			to_chat(user, "<span class='warning'>*BZZZZzzzzzt*</span>")
-	else
-		user.audible_message("<B>[user.GetVoice()]</B>[user.GetAltName()] broadcasts, <FONT size=3>\"[message]\"</FONT>")
+			user.audible_message("<B>[user]</B> broadcasts, <FONT size=3>\"[message]\"</FONT>")
 
-/obj/item/device/megaphone/attack_self(var/mob/living/user)
+		spamcheck = 1
+		spawn(20)
+			spamcheck = 0
+		return
+
+/obj/item/device/megaphone/attack_self(mob/living/user as mob)
+	if(!can_broadcast(user))
+		return
+
 	var/message = sanitize(input(user, "Shout a message?", "Megaphone", null)  as text)
 	if(!message)
 		return
 	message = capitalize(message)
 
-	if(!can_broadcast(user))
-		return
-
-	spamcheck = world.time + 20
 	do_broadcast(user, message)
 
 /obj/item/device/megaphone/emag_act(var/remaining_charges, var/mob/user)
 	if(!emagged)
-		to_chat(user, "<span class='warning'>You overload [src]'s voice synthesizer.</span>")
-		emagged = TRUE
+		to_chat(user, "<span class='warning'>You overload \the [src]'s voice synthesizer.</span>")
+		emagged = 1
 		insults = rand(1, 3)//to prevent caps spam.
-		return TRUE
+		return 1
 
 /obj/item/device/megaphone/super
 	name = "gigaphone"
@@ -129,35 +130,41 @@
 		broadcast_color = new_color
 
 /obj/item/device/megaphone/super/do_broadcast(var/mob/living/user, var/message)
-	if(emagged)
-		if(insults)
-			user.audible_message("<B>[user.GetVoice()]</B>[user.GetAltName()] broadcasts, <FONT size=[broadcast_size] face='[broadcast_font]' color='[broadcast_color]'>\"[pick(insultmsg)]\"</FONT>")
-			if(broadcast_size >= 11)
-				var/turf/T = get_turf(user)
-				playsound(T, 'sound/items/AirHorn.ogg', 100, 1)
-				for(var/mob/living/carbon/M in oviewers(4, T))
-					if(M.get_ear_protection() >= 2)
-						continue
-					M.sleeping = 0
-					M.stuttering += 20
-					M.ear_deaf += 30
-					M.Weaken(3)
-					if(prob(30))
-						M.Stun(10)
-						M.Paralyse(4)
-					else
-						M.make_jittery(50)
-			insults--
+	if ((src.loc == user && usr.stat == 0))
+		if(emagged)
+			if(insults)
+				user.audible_message("<B>[user]</B> broadcasts, <FONT size=[broadcast_size] face='[broadcast_font]' color='[broadcast_color]'>\"[pick(insultmsg)]\"</FONT>")
+				if(broadcast_size >= 11)
+					var/turf/T = get_turf(user)
+					playsound(T, 'sound/items/AirHorn.ogg', 100, 1)
+					for(var/mob/living/carbon/M in oviewers(4, T))
+						if(M.get_ear_protection() >= 2)
+							continue
+						M.sleeping = 0
+						M.stuttering += 20
+						M.ear_deaf += 30
+						M.Weaken(3)
+						if(prob(30))
+							M.Stun(10)
+							M.Paralyse(4)
+						else
+							M.make_jittery(50)
+				insults--
+			else
+				user.audible_message("<span class='critical'>*BZZZZzzzzzt*</span>")
+				if(prob(40) && insults <= 0)
+					var/datum/effect/effect/system/spark_spread/s = new /datum/effect/effect/system/spark_spread
+					s.set_up(2, 1, get_turf(user))
+					s.start()
+					user.visible_message("<span class='warning'>\The [src] sparks violently!</span>")
+					spawn(30)
+						explosion(get_turf(src), -1, -1, 1, 3, adminlog = 1)
+						qdel(src)
+						return
 		else
-			user.audible_message("<span class='critical'>*BZZZZzzzzzt*</span>")
-			if(prob(40) && insults <= 0)
-				var/datum/effect/effect/system/spark_spread/s = new /datum/effect/effect/system/spark_spread
-				s.set_up(2, 1, get_turf(user))
-				s.start()
-				user.visible_message("<span class='warning'>\The [src] sparks violently!</span>")
-				spawn(30)
-					explosion(get_turf(src), -1, -1, 1, 3, adminlog = 1)
-					qdel(src)
-					return
-	else
-		user.audible_message("<B>[user.GetVoice()]</B>[user.GetAltName()] broadcasts, <FONT size=[broadcast_size] face='[broadcast_font]' color='[broadcast_color]'>\"[message]\"</FONT>")
+			user.audible_message("<B>[user]</B> broadcasts, <FONT size=[broadcast_size] face='[broadcast_font]' color='[broadcast_color]'>\"[message]\"</FONT>")
+
+		spamcheck = 1
+		spawn(20)
+			spamcheck = 0
+		return
